+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * New cop `SpaceAfterMethodName` keeps track of method definitions with a space between the method name and the opening parenthesis.
 * New cop `ParenthesesAsGroupedExpression` keeps track of method calls with a space before the opening parenthesis.
 * New Rails cop `HasAndBelongsToMany` checks for uses of `has_and_belongs_to_many`.
+* New Rails cop `ReadAttribute` tracks uses of `read_attribute`.
 
 ### Bugs fixed